requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.0.post1
